requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- end of template --------- #
python-box
pyyaml
ray
opencv-python
accelerate
Pillow
torchvision
pandas
joblib
transformers
peft
datasets
yt-dlp[default]
motion-vector-extractor
termplotlib
matplotlib
scikit-learn
ipykernel